added equal operator and fixed if statement else-jumps being entirely off

+++ src/main/java/org/crayne/mi/bytecode/writer/ByteLoopBound.java
@@ -2,4 +2,4 @@ package org.crayne.mi.bytecode.writer;
 
 import org.crayne.mi.parsing.ast.Node;
 
-public record ByteLoopBound(long beginLabel, long beforeJumpIfLabel, Node forloopInstr) { }
+public record ByteLoopBound(int beginLabel, int beforeJumpIfLabel, Node forloopInstr) { }
